doc: update home page
