+++ docs/src/activities.py
@@ -1,0 +1,6 @@
+from temporalio import activity
+
+
+@activity.defn(name="hello")
+def hello(person: str) -> str:
+    return f"Hello {person}"
