--- docs/src/hello_world.py
@@ -1,8 +1,0 @@
-from icij_worker import AsyncApp
-
-app = AsyncApp("some-app")
-
-
-@app.task
-def hello_world() -> str:
-    return "Hello world"
